refactor: Extract common process monitoring logic into shared library
